Move MessageContent into separate file
