+++ src/messages/content.rs
@@ -1,0 +1,57 @@
+// Copyright 2020 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+
+use crate::{
+    chain::{EldersInfo, GenesisPfxInfo},
+    relocation::RelocateDetails,
+    xor_space::{Prefix, XorName},
+};
+use std::fmt::{self, Debug, Formatter};
+
+#[derive(Eq, PartialEq, Clone, Hash, Serialize, Deserialize)]
+// FIXME - See https://maidsafe.atlassian.net/browse/MAID-2026 for info on removing this exclusion.
+#[allow(clippy::large_enum_variant)]
+/// Content
+pub enum MessageContent {
+    /// Inform neighbours about our new section.
+    NeighbourInfo(EldersInfo),
+    /// User-facing message
+    UserMessage(Vec<u8>),
+    /// Approves the joining node as a routing node.
+    ///
+    /// Sent from Group Y to the joining node.
+    NodeApproval(GenesisPfxInfo),
+    /// Acknowledgement of a consensused section info.
+    AckMessage {
+        /// The prefix of our section when we acknowledge their EldersInfo of version ack_version.
+        src_prefix: Prefix<XorName>,
+        /// The version acknowledged.
+        ack_version: u64,
+    },
+    /// Update sent to Adults and Infants by Elders
+    GenesisUpdate(GenesisPfxInfo),
+    /// Send to a node being relocated from its own section.
+    Relocate(Box<RelocateDetails>),
+}
+
+impl Debug for MessageContent {
+    fn fmt(&self, formatter: &mut Formatter) -> fmt::Result {
+        use self::MessageContent::*;
+        match self {
+            NeighbourInfo(info) => write!(formatter, "NeighbourInfo({:?})", info),
+            UserMessage(content) => write!(formatter, "UserMessage({:?})", content,),
+            NodeApproval(gen_info) => write!(formatter, "NodeApproval({:?})", gen_info),
+            AckMessage {
+                src_prefix,
+                ack_version,
+            } => write!(formatter, "AckMessage({:?}, {})", src_prefix, ack_version),
+            GenesisUpdate(info) => write!(formatter, "GenesisUpdate({:?})", info),
+            Relocate(payload) => write!(formatter, "Relocate({:?})", payload),
+        }
+    }
+}
